Moved workflow task definitions into separate module for reuse.
