+++ example_runtime_config.toml
@@ -12,6 +12,7 @@ checkpoint_mode = 'task_exit'
 # The path to the staging directory
 # e.g. "/gscratch/dirac/kbmod/workflow/staging"
 staging_directory = "/home/drew/code/kbmod-wf/dev_staging"
+output_directory = "/home/drew/code/kbmod-wf/dev_staging/single_chip_workflow"
 
 
 [apps.uri_to_ic]
